+++ tests/test_formatter/test_formatter_output.py
@@ -36,7 +36,7 @@ def _safe_output(output: str) -> str:
     """
     assert pkg_version, 'Looks like version is broken'
 
-    current_version_url = WemakeFormatter._doc_url  # noqa: WPS437
+    current_version_url = WemakeFormatter._doc_url  # noqa: SLF001
     general_version_url = current_version_url.replace(pkg_version, 'xx.xx')
     return output.replace(current_version_url, general_version_url)